refactor: extract shared cde_harvester.core (schemas, db, observability, config, errors)
Single home for code previously duplicated across harvester and db-loader:
- core/db.py: the DB connection string built identically in dataset_state,
  loader and populate_vernaculars
- core/observability.py: Sentry init (2 copies), the get_run_logger fallback
  (3 copies), setup_logging/cleanup_old_logs
- core/schemas.py: Pandera CSV contract + the loader's array dtype maps,
  so the harvester->loader column contract has one home
- core/errors.py: harvest reason codes (moved from harvest_errors.py)
- core/config.py: harvest-config load/resolve helpers (from __main__ and
  prefect_pipeline)

Behavior unchanged; import paths only.
