+++ harvester/cde_harvester/CDEComplianceChecker.py
@@ -1,4 +1,4 @@
-from cde_harvester.harvest_errors import (
+from cde_harvester.core.errors import (
     DEPTH_AND_ALTITUDE,
     INGEST_FLAG_FALSE,
     MISSING_REQUIRED_VARS,
