+++ harvester/tests/unit/test_compliance_checker.py
@@ -15,7 +15,7 @@ from conftest import (
     build_mock_dataset,
 )
 from cde_harvester.CDEComplianceChecker import CDEComplianceChecker
-from cde_harvester.harvest_errors import (
+from cde_harvester.core.errors import (
     DEPTH_AND_ALTITUDE,
     INGEST_FLAG_FALSE,
     MISSING_REQUIRED_VARS,
